Fixed Logical Errors

+++ BinToDec.cpp
@@ -8,8 +8,8 @@ int _tmain(int argc, _TCHAR* argv[])
 {
     const unsigned int nMaxBinDigits = static_cast<unsigned int>(ceil(log2(UINT32_MAX)));
     const std::regex oRegEx("^[0-1]+$");
-    unsigned int nBinPower = 1;
-    unsigned int nBinNum = 0;
+    long long nBinPower = 1L;
+    long long nBinNum = 0L;
     std::string strLine;
     system("cls");
     std::cout << "Input a binary number\r\n";
